Ensure the weight columns are written as Float32 for CASA compatibility

+++ src/corruption/addcorruptions.jl
@@ -8,7 +8,7 @@ include(joinpath("stationgains.jl"))
 include(joinpath("bandpass.jl"))
 include(joinpath("noise.jl"))
 
-function computetotalrms(totalrmsspec::Array{Float64, 4}, totalwtspec::Array{Float64, 4}, obs::CjlObservation)
+function computetotalrms(totalrmsspec::Array{Float32, 4}, totalwtspec::Array{Float32, 4}, obs::CjlObservation)
     """
     Compute total rms (sigma) values from independent noise sources
     """
@@ -58,18 +58,18 @@ function addcorruptions(obs::CjlObservation)
 
     # compute weight columns
     # TODO -- check if the *_SPECTRUM columns are compatible in "manual" vs "uvfits" modes
-    totalrmsspec = ones(Float64, 2, 2, obs.numchan, size(obs.data)[4])
-    totalwtspec = ones(Float64, 2, 2, obs.numchan, size(obs.data)[4])
+    totalrmsspec = ones(Float32, 2, 2, obs.numchan, size(obs.data)[4])
+    totalwtspec = ones(Float32, 2, 2, obs.numchan, size(obs.data)[4])
     computetotalrms(totalrmsspec, totalwtspec, obs)
 
     # convert the sigma_spec, weight_spec and data arrays to the format required by Casacore.jl
     revtotalrmsspec3dres = permutedims(totalrmsspec, (2,1,3,4))
     revtotalrmsspec3d = reshape(revtotalrmsspec3dres, 4, size(revtotalrmsspec3dres)[3], :)
-    revtotalrmsspec = [Matrix{ComplexF32}(revtotalrmsspec3d[:,:,i]) for i in 1:size(revtotalrmsspec3d)[3]]
+    revtotalrmsspec = [Matrix{Float32}(revtotalrmsspec3d[:,:,i]) for i in 1:size(revtotalrmsspec3d)[3]]
 
     revtotalwtspec3dres = permutedims(totalwtspec, (2,1,3,4))
     revtotalwtspec3d = reshape(revtotalwtspec3dres, 4, size(revtotalwtspec3dres)[3], :)
-    revtotalwtspec = [Matrix{ComplexF32}(revtotalwtspec3d[:,:,i]) for i in 1:size(revtotalwtspec3d)[3]]
+    revtotalwtspec = [Matrix{Float32}(revtotalwtspec3d[:,:,i]) for i in 1:size(revtotalwtspec3d)[3]]
 
     revdata3dres = permutedims(obs.data, (2,1,3,4))
     revdata3d = reshape(revdata3dres, 4, size(revdata3dres)[3], :)
